[REFACT] Moved params info to separate files
